+++ app/src/main/java/woowacourse/movie/presentation/model/MovieBookingInfo.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.presentation.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
